Moving length info type into the string table
now we store length info type (8 bit length or 32 bit) in the most
significant bit, rather than in the vector table.  High bit set means 32
bit length (so max 2GB strings), and short lengths are only up to 127.
Not a bad tradeoff though (as it allows for full 32 bit offsets for the
vectors, rather than burning 2 bits).  Went back to 0 specifying when we
use the vector value as well.