using frf_malloc for frf_maker_add
Switched to using frf_malloc for all transient memory used in frf_add
(I.e. vectors and cells).  Saves on destroy the lists later on as well
as having to allocate memory through malloc in the first place.
frf_malloc is always going to be faster.

Switched all the callocs to mallocs as well (to save on initializing
memory) and explicitly initailized the pointers that weren't being
written.